AST: Split off Effects.cpp from Decl.cpp
