+++ lib/AST/CMakeLists.txt
@@ -44,6 +44,7 @@ add_swift_host_library(swiftAST STATIC
   DiagnosticEngine.cpp
   DiagnosticList.cpp
   DocComment.cpp
+  Effects.cpp
   Evaluator.cpp
   Expr.cpp
   ExtInfo.cpp
